pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-ai"
version = "0.1.0"
description = "CLI personal AI agent with memory and tools"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "playwright>=1.40.0",
    "html2text>=2024.2.26",
]

[project.scripts]
personal-ai = "personal_ai.main:main"

[tool.setuptools.packages.find]
where = ["src"]
